requirements.txt
APScheduler==3.6.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
dateparser==1.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-binance==0.7.9
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
service-identity==21.1.0
six==1.16.0
tornado==6.1
tradingview-ta==3.2.3
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
zope.interface==5.4.0
